Move to github servers from code aurora: Update Hashes
Originally, the idea was to keep the hashes the same, but those
hashes refer to recipes which also pull from Code Aurora.  With
that server down, the repo sync might succeed, but the build may
fail.  Fix this by syncing the hashes of the new server with
the hashes from NXP's github server to fetch code from the proper
servers