imx-6.12.20-4.0.0-RC7: Fix 2GB -12 Variant
The focus was on the 4GB variant, due to lack of available hardware.
Since the 4GB can be configured as 2GB, make the various configs in
U-Boot support the 2GB with and without FSPI support as well as address
some issues with ATF which needed a single git revert to make it work
on the 2GB variant of Mini.

Signed-off-by: Adam Ford <aford@beaconembedded.com>